+++ train_dataset.py
@@ -10,7 +10,7 @@ from PIL import Image, ImageDraw
 from torchvision import transforms
 from torch.utils.data import Dataset
 
-import utils
+import src.utils as utils
 
 ALLMASKTYPES = ['single_bbox', 'bbox', 'free_form']
